feat: split MainForm and create Input component
