+++ src/client/components/index.ts
@@ -1,1 +1,1 @@
-export * from './Input'
+export * from './MainForm'
